Stop using SDLx::App, as we don't need it
We don't really need any of its functionality, and for some reason it
causes our window to be resized to a maximized fully black window on exit,
which then stays in place for 2 seconds.

The 2 second delay on exit is caused by SDL_mixer's Mix_Close() call
taking quite long. This is is issue outside of pangzero which I've put on
my todo, but the ugly maximizing of the window, rather then staying as is,
during those 2 seconds is somehow caused by pangzero's use of SDLx::App

Signed-off-by: Hans de Goede <hdegoede@redhat.com>